requirements.txt
aioredis==1.2.0
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.5.1
Automat==0.7.0
Babel==2.6.0
certifi==2019.3.9
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
daphne==2.3.0
defusedxml==0.5.0
Django==2.0.13
django-autoslug-iplweb==1.9.4.dev0
django-countries==5.3.3
django-files-widget==0.2.4.dev0
django-mathfilters==0.4.0
django-phonenumber-field==2.2.0
django-smart-selects==1.5.4
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
djangorestframework-stubs==0.3.0
easy-thumbnails==2.6
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
livereload==2.6.0
msgpack==0.6.1
mysqlclient==1.4.2.post1
oauthlib==3.0.1
phonenumbers==8.10.7
pillow==5.4.1
progressbar2==3.39.3
PyHamcrest==1.9.0
PyJWT==1.7.1
pypiwin32==223
python-utils==2.3.0
python3-openid==3.1.0
pytz==2018.7
pywin32==224
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
tornado==6.0.2
Twisted==19.2.0
txaio==18.8.1
Unidecode==1.0.23
Unipath==1.1
urllib3==1.24.1
virtualenv==16.1.0
virtualenvwrapper-win==1.2.5
zope.interface==4.6.0
